# docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: WOVSDH Nails Booking API
  description: |
    Complete RESTful API for online nail salon booking system with Telegram notifications, Google Calendar integration, and email confirmations.

    ## Base URLs

    - **Local Development**: `http://localhost:3000`
    - **Production**: `https://your-domain.com`

    ## Authentication

    ### Public Endpoints
    Most endpoints are public and don't require authentication. However, Supabase RLS (Row Level Security) ensures data isolation.

    ### Admin Endpoints
    Admin endpoints require `X-Admin-Secret` header with the value of `ADMIN_SECRET_KEY` environment variable.

    ### Internal Endpoints
    Internal endpoints (confirmed bookings, email sending) require `X-Internal-Secret` header with the value of `ADMIN_SECRET_KEY`.

    ## Response Format

    All responses are JSON. Errors follow a standard format:

    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message"
      }
    }
    ```

    ## Dates and Times

    All timestamps are returned in ISO 8601 format with timezone offset (e.g., `2025-06-01T08:00:00+03:00`).

    All dates in query parameters use format `YYYY-MM-DD`.

    ## Rate Limiting

    No rate limiting is currently implemented. For production use, implement rate limiting via Vercel's edge functions or external service.

    ## Webhooks

...
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/telegram/webhook:
    post:
      summary: Telegram Bot webhook endpoint
      description: |
        Receive and process updates from Telegram Bot API. Telegram sends updates as POST requests to this endpoint.

        This endpoint:
        1. Validates the X-Telegram-Bot-Api-Secret-Hash header
        2. Processes callback queries (button clicks)
        3. Handles bot commands from authorized users
        4. Returns 200 OK to all requests (even invalid ones, to avoid redelivery)

        No authentication header required - Telegram includes a cryptographic signature.
      operationId: telegramWebhook
      tags:
        - Telegram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramUpdate'
      responses:
        '200':
          description: Update processed (always returns 200)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
...
components:
  schemas:
...
    TelegramUpdate:
      type: object
      description: Telegram Bot API update
      properties:
        update_id:
          type: integer
        callback_query:
          type: object
          properties:
            id:
              type: string
            from:
              type: object
              properties:
                id:
                  type: integer
                username:
                  type: string
            message:
              type: object
              properties:
                message_id:
                  type: integer
                chat:
                  type: object
                  properties:
                    id:
                      type: integer
            data:
              type: string
              description: Button callback data (format - confirm:booking_id or cancel:booking_id)
        message:
          type: object
          properties:
            message_id:
              type: integer
            from:
              type: object
              properties:
                id:
                  type: integer
            chat:
              type: object
              properties:
                id:
                  type: integer
            text:
              type: string
  securitySchemes:
    AdminSecret:
      type: apiKey
      in: header
      name: X-Admin-Secret
      description: Admin secret key (value of ADMIN_SECRET_KEY env var)

    InternalSecret:
      type: apiKey
      in: header
      name: X-Internal-Secret
      description: Internal secret key (value of ADMIN_SECRET_KEY env var)
...